+++ examples/turnstile.py
@@ -11,5 +11,4 @@ class Turnstile(FSM):
 Turnstile('ts', desc="example turnstile")
 
 if __name__ == '__main__':
-    print("Documentation: https://edzed.readthedocs.io/en/latest/intro.html#cli-demo-tool\n")
     asyncio.run(demo.run_demo())