+++ examples/turnstile2.py
@@ -27,5 +27,4 @@ Turnstile(
 )
 
 if __name__ == '__main__':
-    print("Documentation: https://edzed.readthedocs.io/en/latest/intro.html#cli-demo-tool\n")
     asyncio.run(demo.run_demo())